Extract ConnectionArrayExpander
Step 4 of issue #3066, second half. No intended change in behavior.

The expansion of connection arrays, Connection_Pattern and Connection_Set
including connection replication was about 870 lines of private methods in
InstantiateModel, roughly a third of the file and by far its largest
concern. It moves unchanged into a class of its own, which is constructed
per instance model root where processConnections used to be called and
depends on nothing but the error manager and the progress monitor.

The three Communication_Properties name constants and the two predicates
over them move with it, because the pair of them is what ties the
properties cached on the provisional connection instances to the
properties the expansion reads back. InstantiateModel keeps the caching
and calls the now package-visible isStructuralConnectionProperty to decide
which property definitions to cache early.

Also drops the import of RollbackException, which the javadoc correction
in step 2 already left without a reference.
