+++ core/org.osate.aadl2.instantiation/src/org/osate/aadl2/instantiation/InstantiateModel.java
@@ -49,14 +49,12 @@ import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.ResourceSet;
 import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl;
 import org.eclipse.emf.ecore.util.EcoreUtil;
-import org.eclipse.emf.transaction.RollbackException;
 import org.osate.aadl2.AbstractFeature;
 import org.osate.aadl2.Access;
 import org.osate.aadl2.AccessSpecification;
 import org.osate.aadl2.AccessType;
 import org.osate.aadl2.ArrayDimension;
 import org.osate.aadl2.ArraySize;
-import org.osate.aadl2.BasicPropertyAssociation;
 import org.osate.aadl2.Classifier;
 import org.osate.aadl2.ComponentCategory;
 import org.osate.aadl2.ComponentClassifier;
@@ -67,7 +65,6 @@ import org.osate.aadl2.Context;
 import org.osate.aadl2.DirectedFeature;
 import org.osate.aadl2.DirectionType;
 import org.osate.aadl2.Element;
-import org.osate.aadl2.EnumerationLiteral;
 import org.osate.aadl2.Feature;
 import org.osate.aadl2.FeatureClassifier;
 import org.osate.aadl2.FeatureGroup;
@@ -77,27 +74,20 @@ import org.osate.aadl2.FeaturePrototypeActual;
 import org.osate.aadl2.FeatureType;
 import org.osate.aadl2.FlowEnd;
 import org.osate.aadl2.FlowSpecification;
-import org.osate.aadl2.IntegerLiteral;
-import org.osate.aadl2.ListValue;
 import org.osate.aadl2.ModalElement;
 import org.osate.aadl2.Mode;
 import org.osate.aadl2.ModeBinding;
 import org.osate.aadl2.ModeTransition;
 import org.osate.aadl2.ModeTransitionTrigger;
-import org.osate.aadl2.NamedValue;
 import org.osate.aadl2.Port;
 import org.osate.aadl2.PortSpecification;
 import org.osate.aadl2.Property;
 import org.osate.aadl2.PropertyAssociation;
-import org.osate.aadl2.PropertyExpression;
-import org.osate.aadl2.PropertySet;
 import org.osate.aadl2.Prototype;
-import org.osate.aadl2.RecordValue;
 import org.osate.aadl2.Subcomponent;
 import org.osate.aadl2.TriggerPort;
 import org.osate.aadl2.instance.ComponentInstance;
 import org.osate.aadl2.instance.ConnectionInstance;
-import org.osate.aadl2.instance.ConnectionInstanceEnd;
 import org.osate.aadl2.instance.ConnectionReference;
 import org.osate.aadl2.instance.FeatureCategory;
 import org.osate.aadl2.instance.FeatureInstance;
@@ -117,8 +107,6 @@ import org.osate.aadl2.modelsupport.errorreporting.MarkerAnalysisErrorReporter;
 import org.osate.aadl2.modelsupport.errorreporting.QueuingAnalysisErrorReporter;
 import org.osate.aadl2.modelsupport.modeltraversal.TraverseWorkspace;
 import org.osate.aadl2.modelsupport.resources.OsateResourceUtil;
-import org.osate.aadl2.modelsupport.util.AadlUtil;
-import org.osate.aadl2.util.Aadl2InstanceUtil;
 import org.osate.aadl2.util.Aadl2Util;
 import org.osate.core.OsateCorePlugin;
 import org.osgi.service.prefs.Preferences;
@@ -145,11 +133,6 @@ public class InstantiateModel {
 
 	/* The name for the single mode of a non-modal system */
 	public static final String NORMAL_SOM_NAME = "No Modes";
-
-	/* The properties that determine how a connection is expanded into connection instances */
-	private static final String COMMUNICATION_PROPERTIES = "Communication_Properties";
-	private static final String CONNECTION_PATTERN = "Connection_Pattern";
-	private static final String CONNECTION_SET = "Connection_Set";
 
 	protected AnalysisErrorReporterManager errManager;
 	protected final IProgressMonitor monitor;
@@ -544,7 +527,7 @@ public class InstantiateModel {
 		final List<Property> structuralProperties = new ArrayList<>();
 		final List<Property> remainingProperties = new ArrayList<>();
 		for (Property property : usedProperties) {
-			if (isStructuralConnectionProperty(property)) {
+			if (ConnectionArrayExpander.isStructuralConnectionProperty(property)) {
 				structuralProperties.add(property);
 			} else {
 				remainingProperties.add(property);
@@ -553,7 +536,7 @@ public class InstantiateModel {
 
 		cacheStructuralConnectionProperties(root, structuralProperties);
 		// handle arrays, connection patterns, and connection sets
-		processConnections(root);
+		new ConnectionArrayExpander(monitor, errManager).processConnections(root);
 		checkCanceled();
 
 		final ValidateConnectionsSwitch vcs = new ValidateConnectionsSwitch(monitor, errManager, classifierCache);
@@ -616,8 +599,8 @@ public class InstantiateModel {
 
 	/**
 	 * Cache {@code Connection_Pattern} and {@code Connection_Set} on the provisional connection
-	 * instances of a root. {@link #processConnections(ComponentInstance)} reads them from there to
-	 * expand the connections into the final connection set.
+	 * instances of a root. {@link ConnectionArrayExpander} reads them from there to expand the
+	 * connections into the final connection set.
 	 * <p>
 	 * This uses the same two switches and the same lookup contexts as full property caching, restricted
 	 * to those two property definitions, so the resolved values are the ones full property caching would
@@ -637,31 +620,6 @@ public class InstantiateModel {
 		new CachePropertyAssociationsSwitch(monitor, errManager, structuralProperties, classifierCache,
 				structuralScProps, mode2som).processPreOrderAll(root);
 		checkCanceled();
-	}
-
-	/**
-	 * Is this one of the properties that determine how many connection instances a connection expands
-	 * into? Uses the same test as {@link #getPA(ConnectionInstance, String)}, which is what reads the
-	 * cached values.
-	 */
-	private static boolean isStructuralConnectionProperty(Property property) {
-		return isCommunicationProperty(property, CONNECTION_PATTERN)
-				|| isCommunicationProperty(property, CONNECTION_SET);
-	}
-
-	/**
-	 * Is this the named property of the standard {@code Communication_Properties} property set? The one
-	 * test is shared by {@link #isStructuralConnectionProperty(Property)}, which decides which
-	 * properties are cached on the provisional connection instances, and
-	 * {@link #getPA(ConnectionInstance, String)}, which reads them from there, so the two cannot drift
-	 * apart.
-	 *
-	 * @param property the property definition to test
-	 * @param name the name of the property to look for
-	 */
-	private static boolean isCommunicationProperty(Property property, String name) {
-		return name.equalsIgnoreCase(property.getName()) && property.getOwner() instanceof PropertySet ps
-				&& COMMUNICATION_PROPERTIES.equalsIgnoreCase(ps.getName());
 	}
 
 	// --------------------------------------------------------------------------------------------
@@ -1374,886 +1332,6 @@ public class InstantiateModel {
 				roots.add(new InstantiationRoot(newInstance));
 			}
 		}
-	}
-
-	// --------------------------------------------------------------------------------------------
-	// Methods for connection sets and connection patterns
-	// --------------------------------------------------------------------------------------------
-
-	private void processConnections(ComponentInstance root) throws InterruptedException {
-		checkCanceled();
-		EList<ComponentInstance> replicateConns = new UniqueEList<>();
-		List<ConnectionInstance> toRemove = new ArrayList<>();
-		EList<ConnectionInstance> connilist = root.getAllConnectionInstances();
-		for (ConnectionInstance conni : connilist) {
-			// track all component instances that contain connection instances
-			replicateConns.add(conni.getComponentInstance());
-
-			PropertyAssociation setPA = getPA(conni, CONNECTION_SET);
-			PropertyAssociation patternPA = getPA(conni, CONNECTION_PATTERN);
-
-			if (setPA == null && patternPA == null) {
-				InstancePath srcPath = analyzePath(conni.getContainingComponentInstance(), conni.getSource());
-				InstancePath dstPath = analyzePath(conni.getContainingComponentInstance(), conni.getDestination());
-				if (srcPath == null || dstPath == null) {
-					continue;
-				}
-				// only a connection with an array element at one of its ends is expanded
-				if (srcPath.throughArrayElement() || dstPath.throughArrayElement()) {
-					if (interpretConnectionPatterns(conni, false, null, 0, srcPath.sizes(), 0, dstPath.sizes(), 0,
-							new ArrayList<>(), new ArrayList<>())) {
-						toRemove.add(conni);
-					}
-				}
-			} else if (patternPA != null) {
-				boolean isOpposite = Aadl2InstanceUtil.isOpposite(conni);
-				EcoreUtil.remove(patternPA);
-				List<PropertyExpression> patterns = ((ListValue) patternPA.getOwnedValues().get(0).getOwnedValue())
-						.getOwnedListElements();
-				boolean pathError = false;
-				for (PropertyExpression pe : patterns) {
-					List<PropertyExpression> pattern = ((ListValue) pe).getOwnedListElements();
-
-					InstancePath srcPath = analyzePath(conni.getContainingComponentInstance(), conni.getSource());
-					InstancePath dstPath = analyzePath(conni.getContainingComponentInstance(), conni.getDestination());
-					if (srcPath == null || dstPath == null) {
-						pathError = true;
-						break;
-					}
-					if (!srcPath.throughArrayElement() && !dstPath.throughArrayElement()) {
-						errManager.warning(conni,
-								"Connection pattern specified for connection that does not connect array elements.");
-					} else {
-						if (interpretConnectionPatterns(conni, isOpposite, pattern, 0, srcPath.sizes(), 0,
-								dstPath.sizes(), 0, new ArrayList<>(), new ArrayList<>())) {
-							toRemove.add(conni);
-						}
-					}
-				}
-				if (pathError) {
-					continue;
-				}
-			}
-			// no else as we want both the pattern and the connection set evaluated
-			if (setPA != null) {
-				EcoreUtil.remove(setPA);
-				// TODO-LW: modal conn set allowed?
-				List<Long> srcIndices;
-				List<Long> dstIndices;
-				for (PropertyExpression pe : ((ListValue) setPA.getOwnedValues().get(0).getOwnedValue())
-						.getOwnedListElements()) {
-					RecordValue rv = (RecordValue) pe;
-
-					srcIndices = getIndices(rv, "src");
-					dstIndices = getIndices(rv, "dst");
-					if (Aadl2InstanceUtil.isOpposite(conni)) {
-						// flip indices since we are going in the opposite direction
-						createNewConnection(conni, dstIndices, srcIndices);
-					} else {
-						createNewConnection(conni, srcIndices, dstIndices);
-					}
-				}
-				toRemove.add(conni);
-			}
-		}
-		for (ConnectionInstance conni : toRemove) {
-			EcoreUtil.delete(conni);
-		}
-		replicateConnections(replicateConns);
-	}
-
-	private void replicateConnections(EList<ComponentInstance> replicateConns) {
-		for (ComponentInstance ci : replicateConns) {
-			if (isInArray(ci)) {
-				doReplicateConnections(ci);
-			}
-		}
-	}
-
-	private boolean isInArray(ComponentInstance ci) {
-		while (!(ci instanceof SystemInstance)) {
-			if (!ci.getIndices().isEmpty()) {
-				return true;
-			}
-			ci = ci.getContainingComponentInstance();
-		}
-		return false;
-	}
-
-	private ComponentInstance outermostArray(ComponentInstance ci) {
-		ComponentInstance res = null;
-		while (!(ci instanceof SystemInstance)) {
-			if (!ci.getIndices().isEmpty()) {
-				res = ci;
-			}
-			ci = ci.getContainingComponentInstance();
-		}
-		return res;
-	}
-
-	private void doReplicateConnections(ComponentInstance ci) {
-		String origPath = getComponentInstanceNamePath(ci);
-		ComponentInstance outermostParent = outermostArray(ci).getContainingComponentInstance();
-		doReplicateConnections(ci, origPath, outermostParent);
-	}
-
-	private void doReplicateConnections(ComponentInstance ci, String origPath, ComponentInstance targetParent) {
-		for (ComponentInstance targetci : targetParent.getComponentInstances()) {
-			// do it only for sibling components. Misses out on enclosing arrays
-			if (targetci.getConnectionInstances().isEmpty()) {
-				// only if it does not yet have connection instances
-				String targetpath = getComponentInstanceNamePath(targetci);
-				// compare paths without indices
-				if (origPath.equalsIgnoreCase(targetpath)) {
-					for (ConnectionInstance conni : ci.getConnectionInstances()) {
-						createNewConnection(conni, targetci);
-					}
-				} else if (origPath.startsWith(targetpath)) {
-					doReplicateConnections(ci, origPath, targetci);
-				}
-			}
-		}
-	}
-
-	private String getComponentInstanceNamePath(ComponentInstance ci) {
-		if (ci instanceof SystemInstance) {
-			return ci.getName();
-		}
-		final String path = getComponentInstanceNamePath(ci.getContainingComponentInstance());
-		final String localname = ci.getName();
-
-		return path + "." + localname;
-	}
-
-	private boolean interpretConnectionPatterns(ConnectionInstance conni, boolean isOpposite,
-			List<PropertyExpression> patterns, int offset, List<Integer> srcSizes, int srcOffset,
-			List<Integer> dstSizes, int dstOffset, List<Long> srcIndices, List<Long> dstIndices) {
-		boolean result = true;
-		if (patterns != null ? offset >= patterns.size()
-				: srcOffset == srcSizes.size() && dstOffset == dstSizes.size()) {
-			createNewConnection(conni, srcIndices, dstIndices);
-			return result;
-		}
-		String patternName = "One_to_One";
-		if (patterns == null) {
-			// A default pattern pairs dimensions one-to-one. If only one end has dimensions,
-			// every element on that end maps to the one scalar end.
-			if (srcSizes.isEmpty()) {
-				patternName = isOpposite ? "All_to_One" : "One_To_All";
-			} else if (dstSizes.isEmpty()) {
-				patternName = isOpposite ? "One_To_All" : "All_to_One";
-			}
-		} else {
-			NamedValue nv = (NamedValue) patterns.get(offset);
-			EnumerationLiteral pattern = (EnumerationLiteral) nv.getNamedValue();
-			patternName = pattern.getName();
-		}
-
-		if (!isOpposite && !patternName.equalsIgnoreCase("One_To_All") && (srcOffset >= srcSizes.size())) {
-			errManager.error(conni, "Too few indices for connection source for " + conni.getFullName());
-			return false;
-		}
-		if (!isOpposite && !patternName.equalsIgnoreCase("All_To_One") && (dstOffset >= dstSizes.size())) {
-			errManager.error(conni, "Too few indices for connection destination for " + conni.getFullName());
-			return false;
-		}
-		if (isOpposite && !patternName.equalsIgnoreCase("One_To_All") && (dstOffset >= dstSizes.size())) {
-			errManager.error(conni, "Too few indices for connection source for " + conni.getFullName());
-			return false;
-		}
-		if (isOpposite && !patternName.equalsIgnoreCase("All_To_One") && (srcOffset >= srcSizes.size())) {
-			errManager.error(conni, "Too few indices for connection destination for " + conni.getFullName());
-			return false;
-		}
-		if (patternName.equalsIgnoreCase("All_To_All")) {
-			for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-				srcIndices.add(i);
-				for (long j = 1; j <= dstSizes.get(dstOffset); j++) {
-					dstIndices.add(j);
-					result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-							srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-					dstIndices.remove(dstOffset);
-				}
-				srcIndices.remove(srcOffset);
-			}
-		} else if ((!isOpposite && patternName.equalsIgnoreCase("One_To_All"))
-				|| (isOpposite && patternName.equalsIgnoreCase("All_To_One"))) {
-			for (long j = 1; j <= dstSizes.get(dstOffset); j++) {
-				dstIndices.add(j);
-				result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes, srcOffset,
-						dstSizes, dstOffset + 1, srcIndices, dstIndices);
-				dstIndices.remove(dstOffset);
-			}
-		} else if ((!isOpposite && patternName.equalsIgnoreCase("All_To_One"))
-				|| (isOpposite && patternName.equalsIgnoreCase("One_To_All"))) {
-			for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-				srcIndices.add(i);
-				result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes, srcOffset + 1,
-						dstSizes, dstOffset, srcIndices, dstIndices);
-				srcIndices.remove(srcOffset);
-			}
-		} else {
-			if (!srcSizes.get(srcOffset).equals(dstSizes.get(dstOffset))) {
-				errManager.error(conni,
-						"Array size mismatch (" + patternName + ") on connection " + conni.getFullName() + " in "
-								+ conni.getContainingComponentInstance().getFullName() + ": " + srcSizes.get(srcOffset)
-								+ " at source and " + dstSizes.get(dstOffset) + " at destination.");
-				return false;
-			} else {
-				if (patternName.equalsIgnoreCase("One_To_One")) {
-					for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-						srcIndices.add(i);
-						dstIndices.add(i);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						srcIndices.remove(srcOffset);
-						dstIndices.remove(dstOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Next")) {
-					for (long i = 1; i <= srcSizes.get(srcOffset) - 1; i++) {
-						srcIndices.add(i);
-						dstIndices.add(i + 1);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Previous")) {
-					for (long i = 2; i <= srcSizes.get(srcOffset); i++) {
-						srcIndices.add(i);
-						dstIndices.add(i - 1);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Cyclic_Next")) {
-					for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-						srcIndices.add(i);
-						dstIndices.add(i == srcSizes.get(srcOffset) ? 1 : i + 1);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Cyclic_Previous")) {
-					for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-						srcIndices.add(i);
-						dstIndices.add(i == 1 ? srcSizes.get(srcOffset) : i - 1);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Next_Next")) {
-					for (long i = 1; i <= srcSizes.get(srcOffset) - 2; i++) {
-						srcIndices.add(i);
-						dstIndices.add(i + 2);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Previous_Previous")) {
-					for (long i = 3; i <= srcSizes.get(srcOffset); i++) {
-						srcIndices.add(i);
-						dstIndices.add(i - 2);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Cyclic_Next_Next")) {
-					for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-						srcIndices.add(i);
-						dstIndices
-								.add(i == srcSizes.get(srcOffset) ? 2 : (i == srcSizes.get(srcOffset) - 1 ? 1 : i + 1));
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Cyclic_Previous_Previous")) {
-					for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-						srcIndices.add(i);
-						dstIndices
-								.add(i == 2 ? srcSizes.get(srcOffset) : (i == 1 ? srcSizes.get(srcOffset) - 1 : i - 1));
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Even_To_Even")) {
-					for (long i = 2; i <= srcSizes.get(srcOffset); i = i + 2) {
-						srcIndices.add(i);
-						dstIndices.add(i);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else if (patternName.equalsIgnoreCase("Odd_To_Odd")) {
-					for (long i = 1; i <= srcSizes.get(srcOffset); i = i + 2) {
-						srcIndices.add(i);
-						dstIndices.add(i);
-						result &= interpretConnectionPatterns(conni, isOpposite, patterns, offset + 1, srcSizes,
-								srcOffset + 1, dstSizes, dstOffset + 1, srcIndices, dstIndices);
-						dstIndices.remove(dstOffset);
-						srcIndices.remove(srcOffset);
-					}
-				} else {
-					/*
-					 * A pattern this method does not know expands into nothing. Report it and keep the
-					 * connection: returning the initial true would tell the caller that the connection was
-					 * expanded, and the caller would delete it without a replacement.
-					 */
-					errManager.error(conni, "Unsupported connection pattern '" + patternName + "' on connection "
-							+ conni.getFullName());
-					return false;
-				}
-			}
-		}
-		return result;
-	}
-
-	private List<Long> getIndices(RecordValue rv, String field) {
-		List<Long> indices = new ArrayList<>();
-		for (BasicPropertyAssociation fv : rv.getOwnedFieldValues()) {
-			if (fv.getProperty().getName().equalsIgnoreCase(field)) {
-				ListValue lv = (ListValue) fv.getOwnedValue();
-				EList<PropertyExpression> vlist = lv.getOwnedListElements();
-				for (PropertyExpression elem : vlist) {
-					indices.add(((IntegerLiteral) elem).getValue());
-				}
-			}
-		}
-		return indices;
-	}
-
-	/**
-	 * The association of a {@code Communication_Properties} property cached on a provisional connection
-	 * instance, or {@code null} if the connection has none.
-	 *
-	 * @param conni the connection instance to look in
-	 * @param name the name of the property to look for
-	 */
-	private PropertyAssociation getPA(ConnectionInstance conni, String name) {
-		for (PropertyAssociation pa : conni.getOwnedPropertyAssociations()) {
-			if (isCommunicationProperty(pa.getProperty(), name)) {
-				return pa;
-			}
-		}
-		return null;
-	}
-
-	/**
-	 * Returns the instance objects from {@code end} up to, but excluding, {@code container}. A
-	 * connection end path is relative to its containing component instance, so reaching any other root
-	 * is an error rather than a successful stopping condition.
-	 */
-	private List<InstanceObject> getConnectionEndPath(ComponentInstance container, ConnectionInstanceEnd end) {
-		if (container == null) {
-			String endPath = end == null ? "<null>" : end.getInstanceObjectPath();
-			String message = "Cannot analyze connection end '" + endPath + "' without a containing component instance";
-			if (end == null) {
-				errManager.internalError(message);
-			} else {
-				errManager.error(end, message);
-			}
-			return null;
-		}
-		List<InstanceObject> path = new ArrayList<>();
-		InstanceObject current = end;
-		while (current != container) {
-			if (current == null || !(current.getOwner() instanceof InstanceObject owner)) {
-				String endPath = end == null ? "<null>" : end.getInstanceObjectPath();
-				errManager.error(container, "Connection end '" + endPath + "' is not contained in component instance '"
-						+ container.getInstanceObjectPath() + "'");
-				return null;
-			}
-			path.add(current);
-			current = owner;
-		}
-		return path;
-	}
-
-	/**
-	 * The path from a connection instance end up to its containing component instance, described by one
-	 * entry per step for the names and the dimensions, and by one entry per array dimension for the sizes
-	 * and the indices.
-	 *
-	 * @param names the name of every instance object along the path, bottom up
-	 * @param dims the number of declared array dimensions of every instance object along the path, zero
-	 *            for one that is not an array element
-	 * @param sizes the size of every array dimension along the path, bottom up
-	 * @param indices the index of the path in every array dimension along the path, bottom up
-	 */
-	private record InstancePath(List<String> names, List<Integer> dims, List<Integer> sizes, List<Long> indices) {
-		/**
-		 * Does this path go through an array element? An empty size list and a dimension list of nothing
-		 * but zeros are the same statement, because every step contributes one size per declared
-		 * dimension.
-		 */
-		boolean throughArrayElement() {
-			return !sizes.isEmpty();
-		}
-	}
-
-	/**
-	 * Describe the path from a connection instance end up to the component instance that contains the
-	 * connection instance.
-	 *
-	 * @param container the component instance the path is relative to
-	 * @param end the connection instance end the path starts at
-	 * @return the path, or {@code null} if the end is not contained in {@code container}, which is
-	 *         reported
-	 */
-	private InstancePath analyzePath(ComponentInstance container, ConnectionInstanceEnd end) {
-		List<InstanceObject> path = getConnectionEndPath(container, end);
-		if (path == null) {
-			return null;
-		}
-		List<String> names = new ArrayList<>();
-		List<Integer> dims = new ArrayList<>();
-		List<Integer> sizes = new ArrayList<>();
-		List<Long> indices = new ArrayList<>();
-		for (InstanceObject current : path) {
-			int d = 0;
-
-			names.add(current.getName());
-			if (current instanceof ComponentInstance componentInstance) {
-				d = componentInstance.getSubcomponent().getArrayDimensions().size();
-				if (d != 0) {
-					indices.addAll(componentInstance.getIndices());
-					for (ArrayDimension ad : componentInstance.getSubcomponent().getArrayDimensions()) {
-						ArraySize as = ad.getSize();
-
-						sizes.add((int) getElementCount(as, current.getContainingComponentInstance()));
-					}
-				}
-			} else if (current instanceof FeatureInstance featureInstance && featureInstance.getIndex() != 0) {
-				d = 1;
-				indices.add(featureInstance.getIndex());
-				Feature f = featureInstance.getFeature();
-				ArraySize as = f.getArrayDimensions().get(0).getSize();
-
-				sizes.add((int) getElementCount(as, current.getContainingComponentInstance()));
-			}
-			dims.add(d);
-		}
-		return new InstancePath(names, dims, sizes, indices);
-	}
-
-	/**
-	 * Create a copy of the connection instance with the specified indices for the source and the
-	 * destination.
-	 *
-	 * @param conni the connection instance to copy
-	 * @param srcIndices the indices of the source end
-	 * @param dstIndices the indices of the destination end
-	 */
-	private void createNewConnection(ConnectionInstance conni, List<Long> srcIndices, List<Long> dstIndices) {
-		ComponentInstance container = conni.getContainingComponentInstance();
-		ConnectionInstance newConn = EcoreUtil.copy(conni);
-		newConn.setSource(null);
-		newConn.setDestination(null);
-		ConnectionReference topConnRef = Aadl2InstanceUtil.getTopConnectionReference(newConn);
-		InstancePath srcPath = analyzePath(container, conni.getSource());
-		if (srcPath == null) {
-			return;
-		}
-		if (srcIndices.size() != srcPath.sizes().size() &&
-		// filter out one side being an element without index (array of 1) (many to one mapping)
-				!(srcPath.sizes().isEmpty() && dstIndices.size() == 1)) {
-			errManager.error(container,
-					"Source indices " + srcIndices + " do not match source dimension " + srcPath.sizes().size());
-		}
-		InstanceObject src = resolveConnectionInstancePath(newConn, topConnRef, container, container, srcPath, srcIndices,
-				true);
-		InstancePath dstPath = analyzePath(container, conni.getDestination());
-		if (dstPath == null) {
-			return;
-		}
-		if (dstIndices.size() != dstPath.sizes().size() &&
-		// filter out one side being an element without index (array of 1) (many to one mapping)
-				!(dstPath.sizes().isEmpty() && dstIndices.size() == 1)) {
-			errManager.error(container,
-					"For " + newConn.getConnectionReferences().get(0).getFullName() + " : " + newConn.getFullName()
-							+ ", destination indices " + dstIndices + " do not match destination dimension "
-							+ dstPath.sizes().size());
-		}
-		InstanceObject dst = resolveConnectionInstancePath(newConn, topConnRef, container, container, dstPath, dstIndices,
-				false);
-		if (src == null) {
-			errManager.error(container, "Connection source not found");
-		}
-		if (dst == null) {
-			errManager.error(container, "Connection destination not found");
-		}
-		if (src == null || dst == null) {
-			return;
-		}
-
-		String name = connectionInstanceName(container, src, dst);
-		ConnectionInstance duplicate = (ConnectionInstance) AadlUtil
-				.findNamedElementInList(container.getConnectionInstances(), name);
-		if (duplicate != null && duplicate != conni) { // conni will be removed later
-			errManager.warning(container, "There is already another connection between the same endpoints");
-		}
-		newConn.setSource((ConnectionInstanceEnd) src);
-		newConn.setDestination((ConnectionInstanceEnd) dst);
-		alignConnectionReferenceEndpoints(newConn);
-		newConn.setName(name);
-		container.getConnectionInstances().add(newConn);
-
-	}
-
-	/**
-	 * Create a copy of the connection instance in another element of the array the connection instance
-	 * belongs to.
-	 *
-	 * @param conni the connection instance to copy
-	 * @param targetComponent the component instance the copy belongs to
-	 */
-	private void createNewConnection(ConnectionInstance conni, ComponentInstance targetComponent) {
-		ComponentInstance origComponent = conni.getContainingComponentInstance();
-		ConnectionInstance newConn = EcoreUtil.copy(conni);
-		newConn.setSource(null);
-		newConn.setDestination(null);
-		ConnectionReference topConnRef = Aadl2InstanceUtil.getTopConnectionReference(newConn);
-		InstancePath srcPath = analyzePath(origComponent, conni.getSource());
-		if (srcPath == null) {
-			return;
-		}
-		InstanceObject src = resolveConnectionInstancePath(newConn, topConnRef, targetComponent, targetComponent, srcPath,
-				srcPath.indices(), true);
-		InstancePath dstPath = analyzePath(origComponent, conni.getDestination());
-		if (dstPath == null) {
-			return;
-		}
-		InstanceObject dst = resolveConnectionInstancePath(newConn, topConnRef, targetComponent, targetComponent, dstPath,
-				dstPath.indices(), false);
-		if (src == null) {
-			errManager.error(targetComponent, "Connection source not found");
-		}
-		if (dst == null) {
-			errManager.error(targetComponent, "Connection destination not found");
-		}
-		if (src == null || dst == null) {
-			return;
-		}
-
-		relocateConnectionReferenceContexts(newConn, origComponent, targetComponent);
-		newConn.setSource((ConnectionInstanceEnd) src);
-		newConn.setDestination((ConnectionInstanceEnd) dst);
-		alignConnectionReferenceEndpoints(newConn);
-		newConn.setName(connectionInstanceName(targetComponent, src, dst));
-		targetComponent.getConnectionInstances().add(newConn);
-
-	}
-
-	/**
-	 * The name of a connection instance, which is the path of its source and the path of its destination
-	 * relative to the component instance that contains it.
-	 *
-	 * @param container the component instance that contains the connection instance
-	 * @param src the source of the connection instance
-	 * @param dst the destination of the connection instance
-	 * @return the name of the connection instance
-	 */
-	private static String connectionInstanceName(ComponentInstance container, InstanceObject src, InstanceObject dst) {
-		String containerPath = container.getInstanceObjectPath();
-		return relativeInstanceObjectPath(containerPath, src) + " --> "
-				+ relativeInstanceObjectPath(containerPath, dst);
-	}
-
-	/**
-	 * The path of a connection end relative to the component instance that contains the connection
-	 * instance, or its absolute path if it is not below that component instance.
-	 */
-	private static String relativeInstanceObjectPath(String containerPath, InstanceObject end) {
-		String endPath = end.getInstanceObjectPath();
-		return endPath.startsWith(containerPath) ? endPath.substring(containerPath.length() + 1) : endPath;
-	}
-
-	/**
-	 * Keep the ends of a copied connection's reference chain aligned with the ends that were resolved
-	 * for the copy. Re-resolving from a declaration that names a feature group can otherwise leave the
-	 * reference at the group, and a failed nested lookup can leave it in the original array element.
-	 *
-	 * @param connection the copied connection whose source and destination have been resolved
-	 */
-	private void alignConnectionReferenceEndpoints(ConnectionInstance connection) {
-		List<ConnectionReference> references = connection.getConnectionReferences();
-		if (!references.isEmpty()) {
-			references.getFirst().setSource(connection.getSource());
-			references.getLast().setDestination(connection.getDestination());
-		}
-	}
-
-	/**
-	 * Point the connection references of a replicated connection instance at the array element that now
-	 * contains them. The copy keeps the contexts of the element it was copied from. Resolving the
-	 * endpoints gives every enclosed connection reference a context in the target element, but nothing
-	 * updates the context of the reference that goes across in the copied element.
-	 *
-	 * @param newConn the copy of the connection instance
-	 * @param origComponent the component instance that contains the original connection instance
-	 * @param targetComponent the component instance that will contain the copy
-	 */
-	private void relocateConnectionReferenceContexts(ConnectionInstance newConn, ComponentInstance origComponent,
-			ComponentInstance targetComponent) {
-		for (ConnectionReference connRef : newConn.getConnectionReferences()) {
-			ComponentInstance relocated = relocateComponentInstance(connRef.getContext(), origComponent,
-					targetComponent);
-			if (relocated != null) {
-				connRef.setContext(relocated);
-			}
-		}
-	}
-
-	/**
-	 * Find the component instance in the target element that corresponds to a component instance in the
-	 * element the connection instance was copied from.
-	 *
-	 * @param original the component instance to relocate
-	 * @param origComponent the root of the subtree the original belongs to
-	 * @param targetComponent the root of the subtree to relocate into
-	 * @return the corresponding component instance, or null if the original is neither
-	 *         <code>origComponent</code> nor contained in it, or if it has no counterpart
-	 */
-	private ComponentInstance relocateComponentInstance(ComponentInstance original, ComponentInstance origComponent,
-			ComponentInstance targetComponent) {
-		if (original == null) {
-			return null;
-		}
-		if (original == origComponent) {
-			return targetComponent;
-		}
-		ComponentInstance parent = relocateComponentInstance(original.getContainingComponentInstance(), origComponent,
-				targetComponent);
-		if (parent == null) {
-			return null;
-		}
-		for (ComponentInstance candidate : parent.getComponentInstances()) {
-			if (candidate.getName().equalsIgnoreCase(original.getName())
-					&& candidate.getIndices().equals(original.getIndices())) {
-				return candidate;
-			}
-		}
-		return null;
-	}
-
-	/**
-	 * resolve downgoing source or destination of the connection reference.
-	 * we do so by re-retrieving the feature instance based on the existing connection instance end name.
-	 * If the connection reference is up or down going we also fill in the other end.
-	 * @param targetConnRef the connection reference to resolve
-	 * @param outerConnRef the connection reference one level up
-	 * @param target the component instance to resolve the end in
-	 * @param doSource whether to resolve the source end or the destination end
-	 * @param idx the index of the feature instance to resolve to
-	 * @param fgidx the index of the enclosing feature group instance to resolve to
-	 * @return the resolved end
-	 */
-	private ConnectionInstanceEnd resolveConnectionReference(ConnectionReference targetConnRef,
-			ConnectionReference outerConnRef, ComponentInstance target, boolean doSource, long idx, long fgidx) {
-		ConnectionInstanceEnd src = targetConnRef.getSource();
-		ConnectionInstanceEnd dst = targetConnRef.getDestination();
-		if (doSource) {
-			if (target.getName().equalsIgnoreCase(src.getName())) {
-				// we point to a component instance, such as a bus or data component in an access connection
-				targetConnRef.setSource(target);
-			} else if (src instanceof FeatureInstance) {
-				// re-resolve the source feature
-				ConnectionInstanceEnd found = (ConnectionInstanceEnd) AadlUtil
-						.findNamedElementInList(target.getFeatureInstances(), src.getName(), idx);
-				if (found == null && src.getOwner() instanceof FeatureInstance parent) {
-					found = (ConnectionInstanceEnd) AadlUtil.findNamedElementInList(target.getFeatureInstances(),
-							parent.getName(), fgidx);
-				}
-				if (found != null) {
-					targetConnRef.setSource(found);
-				}
-
-			}
-			// now we need to resolve the upper end (destination)
-			if (targetConnRef != outerConnRef) {
-				// we need to fix the context of the connection reference
-				ConnectionInstanceEnd outerSrc = outerConnRef.getSource();
-				targetConnRef.setContext(outerSrc.getComponentInstance());
-				// we are not at the top so we fix up the upper end of the connection reference
-				if ((dst.getOwner() instanceof ComponentInstance)
-						&& dst.getName().equalsIgnoreCase(outerSrc.getName())) {
-					targetConnRef.setDestination(outerSrc);
-				} else {
-					// the outer source points to the enclosing feature group. reresolve the feature in this feature group
-					ConnectionInstanceEnd found = (ConnectionInstanceEnd) AadlUtil.findNamedElementInList(
-							((FeatureInstance) outerSrc).getFeatureInstances(), dst.getName(), idx);
-					if (found == null && dst.getOwner() instanceof FeatureInstance parent) {
-						found = (ConnectionInstanceEnd) AadlUtil.findNamedElementInList(target.getFeatureInstances(),
-								parent.getName(), fgidx);
-					}
-					if (found != null) {
-						targetConnRef.setDestination(found);
-					}
-				}
-			}
-			return targetConnRef.getSource();
-		} else {
-			if (target.getName().equalsIgnoreCase(dst.getName())) {
-				// we point to a component instance, such as a bus or data component in an access connection
-				targetConnRef.setDestination(target);
-			} else if (dst instanceof FeatureInstance) {
-				// re-resolve the source feature
-				ConnectionInstanceEnd found = (ConnectionInstanceEnd) AadlUtil
-						.findNamedElementInList(target.getFeatureInstances(), dst.getName(), idx);
-				if (found == null && dst.getOwner() instanceof FeatureInstance parent) {
-					found = (ConnectionInstanceEnd) AadlUtil.findNamedElementInList(target.getFeatureInstances(),
-							parent.getName(), fgidx);
-				}
-				if (found != null) {
-					targetConnRef.setDestination(found);
-				}
-			}
-			// now we need to resolve the upper end (source)
-			if ((outerConnRef != null) && (targetConnRef != outerConnRef)) {
-				// we need to fix the context of the connection reference
-				ConnectionInstanceEnd outerDst = outerConnRef.getDestination();
-				targetConnRef.setContext(outerDst.getComponentInstance());
-				// we are not at the top so we fix up the upper end of the connection reference
-				if ((src.getOwner() instanceof ComponentInstance)
-						&& src.getName().equalsIgnoreCase(outerDst.getName())) {
-					targetConnRef.setSource(outerDst);
-				} else {
-					// the outer source points to the enclosing feature group. reresolve the feature in this feature group
-					ConnectionInstanceEnd found = (ConnectionInstanceEnd) AadlUtil.findNamedElementInList(
-							((FeatureInstance) outerDst).getFeatureInstances(), src.getName(), idx);
-					if (found == null && src.getOwner() instanceof FeatureInstance parent) {
-						found = (ConnectionInstanceEnd) AadlUtil.findNamedElementInList(target.getFeatureInstances(),
-								parent.getName(), fgidx);
-					}
-					if (found != null) {
-						targetConnRef.setSource(found);
-					}
-				}
-			}
-			return targetConnRef.getDestination();
-		}
-	}
-
-	/**
-	 * this method resolves the connection instance from the top connection reference down the source or the destination
-	 * @param newconn Connection Instance whose paths need to be resolved
-	 * @param topref Connection Reference going across components
-	 * @param resolutionRoot component instance from which the path is resolved
-	 * @param diagnosticTarget resource-backed object used for diagnostics
-	 * @param path the names and dimensions of the path of the original connection end, bottom up
-	 * @param indices The indices to be used for elements that are arrays
-	 * @param doSource Go down the source path or the destination path
-	 * @return ConnectionInstanceEnd the ultimate source/destination object (feature instance or component instance)
-	 */
-	private ConnectionInstanceEnd resolveConnectionInstancePath(ConnectionInstance newconn, ConnectionReference topref,
-			ComponentInstance resolutionRoot, Element diagnosticTarget, InstancePath path, List<Long> indices,
-			boolean doSource) {
-		final List<String> names = path.names();
-		final List<Integer> dims = path.dims();
-		// the connection reference to be resolved
-		ConnectionReference targetConnRef = topref;
-		ConnectionReference outerConnRef = topref;
-		ConnectionInstanceEnd resolutionContext = resolutionRoot;
-		// we have to process the indices backwards since we go top down
-		// offset starts with the last element of the indices array
-		int offset = indices.size() - 1;
-		int count = dims.size() - 1;
-		ConnectionInstanceEnd result = null;
-		for (int nameidx = names.size() - 1; nameidx >= 0; nameidx--) {
-			String name = names.get(nameidx);
-			List<InstanceObject> owned = new ArrayList<>();
-			int dim = dims.get(count);
-			if (resolutionContext instanceof ComponentInstance componentContext) {
-				// if nextConnRef is null it is because we are going to look up feature instances inside the last component instance
-				owned.addAll(componentContext.getComponentInstances());
-				owned.addAll(componentContext.getFeatureInstances());
-			} else if (resolutionContext instanceof FeatureInstance featureContext) {
-				owned.addAll(featureContext.getFeatureInstances());
-			}
-
-			if (dim == 0) {
-				resolutionContext = (ConnectionInstanceEnd) AadlUtil.findNamedElementInList(owned, name);
-				// targetConnRef could be null once we are at the end and will resolve the feature name(s)
-//				if (targetConnRef != null&& resolutionContext instanceof ComponentInstance){
-//					result = resolveConnectionReference(targetConnRef, outerConnRef,(ComponentInstance)resolutionContext, doSource) ;
-//				} else {
-//					// the resolved feature has been found
-//					result = resolutionContext;
-//				}
-			} else {
-				// find the object based on its name and indices
-				outer: for (InstanceObject io : owned) {
-					if (io.getName().equalsIgnoreCase(name)) {
-						try {
-							if (io instanceof ComponentInstance componentInstance) {
-								// we need to deal with possibly more than one index
-								int d = dim - 1;
-								for (long i : componentInstance.getIndices()) {
-									if (i != indices.get(offset - d)) {
-										continue outer;
-									}
-									d--;
-								}
-							} else {
-								// we have a feature that may have an index or zero index
-								if (((FeatureInstance) io).getIndex() != indices.get(offset)) {
-									continue outer;
-								}
-							}
-						} catch (IndexOutOfBoundsException e) {
-							errManager.warning(diagnosticTarget,
-									"Too few indices for connection end, using first array element");
-						}
-						resolutionContext = (ConnectionInstanceEnd) io;
-						break;
-					}
-				}
-			}
-			if (resolutionContext == null) {
-				return null;
-			}
-			// resolve the connref
-			if (targetConnRef != null && resolutionContext instanceof ComponentInstance) {
-				int dimfeature = dims.get(0);
-				int dimfg = 0;
-				if (dims.size() > 1) {
-					dimfg = dims.get(1);
-				}
-				result = resolveConnectionReference(targetConnRef, outerConnRef, (ComponentInstance) resolutionContext,
-						doSource, dimfeature == 0 ? 0 : indices.get(0),
-						dimfg == 0 ? 0 : (dimfeature == 0 ? indices.get(0) : indices.get(1)));
-			} else {
-				// the resolved feature has been found
-				result = resolutionContext;
-			}
-			if (doSource) {
-				if (targetConnRef != null && result instanceof FeatureInstance) {
-					targetConnRef.setSource(result);
-				}
-				outerConnRef = targetConnRef;
-				targetConnRef = Aadl2InstanceUtil.getPreviousConnectionReference(newconn, outerConnRef);
-			} else {
-				if (targetConnRef != null && result instanceof FeatureInstance) {
-					targetConnRef.setDestination(result);
-				}
-				outerConnRef = targetConnRef;
-				targetConnRef = Aadl2InstanceUtil.getNextConnectionReference(newconn, outerConnRef);
-			}
-			// reduce the offset by the processed indices of the element we just looked up
-			offset -= dim;
-			// reduce the index into the dims array to get the next number of dimensions
-			count--;
-			// now we need to update the connref pointers
-		}
-		return result;
 	}
 
 	// --------------------------------------------------------------------------------------------
